Roles and Permissions, Projects, DB Seed
Added the Spatie package to handle roles and permissions

Added UsersController, PermissionsController, and RolesController

Added views for each object

Added Project object along with its controller, model, and views

Changed Course<->Module relationship from many-to-many to one-to-many

Changed Module<->Lesson relationship from many-to-many to one-to-many

Removed migrations for course_module table and for lesson_module table

Added database seed to initialize db with permissions, roles, and users

Modified readme to explain about database seeding and how to use it

+++ app/Http/Controllers/ExercisesController.php
@@ -56,7 +56,8 @@ class ExercisesController extends Controller
         $exercise->pre_code = $request->input('pre_code');
         $exercise->start_code = $request->input('start_code');
         $exercise->test_code = $request->input('test_code');
-
+        $exercise->user_id = auth()->user()->id;
+        
         $exercise->save();
         
         return redirect(url('/exercises'))->with('success', 'Exercise Created');
@@ -83,6 +84,11 @@ class ExercisesController extends Controller
     public function edit($id)
     {
         $exercise = Exercise::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id != $exercise->user_id){
+            return redirect(url('/exercises'))->with('error', 'Unauthorized Page');
+        }
 
         return view('exercises.edit')->with('exercise', $exercise);
     }
@@ -122,6 +128,11 @@ class ExercisesController extends Controller
     public function destroy($id)
     {
         $exercise = Exercise::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id != $exercise->user_id){
+            return redirect(url('/exercises'))->with('error', 'Unauthorized Page');
+        }
 
         $exercise->delete();
         return redirect(url('/exercises'))->with('success', 'Exercise Deleted');
